% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/greet.R
\name{greet}
\alias{greet}
\title{Greet a user}
\usage{
greet(name = "User")
}
\arguments{
\item{name}{put in whatever you want! character; name of a user}
}
\value{
nothing. unless you create an assignable output.
}
\description{
Prints a greeting fora user.
}
\examples{
greet()
greet("Laura")
}
