% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{mydata}
\alias{mydata}
\title{World Health Organization TB data}
\format{
\subsection{\code{who}}{

A data frame with 150 rows and 5 columns:
\describe{
\item{sepal.length}{length of sepals}
...
}
}
}
\source{
\url{https://www.who.int/teams/global-tuberculosis-programme/data}
}
\usage{
mydata
}
\description{
A subset of data from the World Health Organization Global Tuberculosis
Report ...
}
\keyword{datasets}
